connections.py
import sqlite3
from sqlite3 import Error

import mysql.connector
from mysql.connector import Error


def create_connection_sqlite3(path):
    connection = None
    try:
        connection = sqlite3.connect(path)
        print("Connection to SQLite DB successful")
    except Error as e:
        print(f"The error '{e}' occurred")

    return connection


def create_connection_mysql(host_name, user_name, user_password):
    connection = None
    try:
        connection = mysql.connector.connect(
            host=host_name,
            user=user_name,
            passwd=user_password
        )
        print("Connection to MySQL DB successful")
    except Error as e:
        print(f"The error '{e}' occurred")


if __name__ == '__main__':
    create_connection_sqlite3('.\\db\\wrk.sqlite')
