requirements.txt
mysql-connector-python-8.0.30 protobuf-3.20.1;
psycopg2;